requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via python-binance
aiosignal==1.4.0
    # via aiohttp
altair==6.0.0
    # via
    #   -r requirements.in
    #   streamlit
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backtesting==0.6.5
    # via -r requirements.in
bcrypt==5.0.0
    # via streamlit-authenticator
beautifulsoup4==4.14.3
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
bokeh==3.8.2
    # via backtesting
cachetools==6.2.6
    # via streamlit
captcha==0.7.1
    # via streamlit-authenticator
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
contourpy==1.3.3
    # via bokeh
cryptography==46.0.4
    # via streamlit-authenticator
dateparser==1.2.2
    # via python-binance
extra-streamlit-components==0.1.81
    # via streamlit-authenticator
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
greenlet==3.3.1
    # via playwright
idna==3.11
    # via
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   altair
    #   bokeh
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
millify==0.1.1
    # via -r requirements.in
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
narwhals==2.15.0
    # via
    #   altair
    #   bokeh
    #   plotly
numpy==2.4.1
    # via
    #   -r requirements.in
    #   backtesting
    #   bokeh
    #   contourpy
    #   pandas
    #   pydeck
    #   streamlit
    #   ta
packaging==26.0
    # via
    #   altair
    #   bokeh
    #   plotly
    #   streamlit
pandas==2.3.3
    # via
    #   -r requirements.in
    #   backtesting
    #   bokeh
    #   streamlit
    #   ta
pillow==12.1.0
    # via
    #   bokeh
    #   captcha
    #   streamlit
playwright==1.57.0
    # via -r requirements.in
plotly==6.5.2
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.4
    # via streamlit
pyarrow==23.0.0
    # via streamlit
pycparser==3.0
    # via cffi
pycryptodome==3.23.0
    # via python-binance
pydeck==0.9.1
    # via streamlit
pyee==13.0.0
    # via playwright
pyjwt==2.10.1
    # via streamlit-authenticator
python-binance==1.0.34
    # via -r requirements.in
python-crontab==3.3.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   dateparser
    #   pandas
pytz==2025.2
    # via
    #   -r requirements.in
    #   dateparser
    #   pandas
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   bokeh
    #   streamlit-authenticator
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.1.15
    # via dateparser
requests==2.32.5
    # via
    #   -r requirements.in
    #   python-binance
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
schedule==1.2.2
    # via -r requirements.in
scheduler==0.8.8
    # via -r requirements.in
six==1.17.0
    # via
    #   python-binance
    #   python-dateutil
smmap==5.0.2
    # via gitdb
soupsieve==2.8.3
    # via beautifulsoup4
streamlit==1.52.2
    # via
    #   -r requirements.in
    #   extra-streamlit-components
    #   streamlit-authenticator
streamlit-authenticator==0.4.2
    # via -r requirements.in
ta==0.11.0
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via
    #   bokeh
    #   streamlit
typeguard==4.4.4
    # via scheduler
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   altair
    #   beautifulsoup4
    #   pyee
    #   referencing
    #   streamlit
    #   typeguard
tzdata==2025.3
    # via pandas
tzlocal==5.3.1
    # via dateparser
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via streamlit
websockets==16.0
    # via python-binance
xyzservices==2025.11.0
    # via bokeh
yarl==1.22.0
    # via aiohttp
